requirements.txt
streamlit
sqlalchemy
pymysql
werkzeug
flask
flask_mail
langchain_openai
pydantic
python-dotenv